requirements.txt
aiofiles==24.1.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.5
decorator==5.1.1
etelemetry==0.3.1
executing==2.1.0
fastai==2.7.17
fastapi==0.115.0
fastcore==1.7.8
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
fitz==0.0.1.dev2
fonttools==4.53.1
frontend==0.0.3
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langcodes==3.4.0
language_data==1.2.0
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
narwhals==1.8.2
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
poppler-utils==0.1.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.28.2
prov==2.0.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.5
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.3.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
